test: restore QueueChecker integration suite alongside unit tests
Keep mock-based QueueCheckerUnitTest for config/age edge cases, and put
back IntegrationTestCase coverage (real messenger_messages + QueryBuilder)
including failed-queue exclusion, allowlist, and per-queue grace.
